==> notebooks/3_generate_negative_data.ipynb <==
%cd ..

from scripts.utils_latest import *


import pandas as pd
import matplotlib.pyplot as plt

import matplotlib.colors as mcolors
import seaborn as sns
pd.set_option('display.max_columns', None)


df = pd.read_csv("results/1_generate_data.csv")
clash_df = clash()
df.head()

negative_df = df[["mre_region", "mirna_sequence", "mirna_accession"]]

# shuffle mirna sequences using levenstein distance
for i, row in negative_df.iterrows():
    original_string = row['mirna_sequence']
    # exclude the original string from the column
    column = negative_df['mirna_sequence'].drop(i)
    new_string = find_most_different_string(original_string, column)
    negative_df.at[i, 'shuffled_sequence'] = new_string

# renaming cols for convention
rename_dict = {"mrna_sequence": "mre_region", "mirna_sequence": "old_mirna_sequence",
               "mirna_accession": "old_mirna_accession", "shuffled_sequence": "mirna_sequence"}
negative_df = negative_df.rename(columns=rename_dict)

# adding accessions for new mirna sequences
accession_dict = clash_df.set_index('mirna_sequence')[
    'mirna_accession'].to_dict()
negative_df['mirna_accession'] = negative_df['mirna_sequence'].map(
    accession_dict)

negative_df = negative_df.rename(columns={"mre_region": "mrna_sequence"})


# finding matches on negative data
negative_df_results = find_matches_with_rnaduplex(negative_df)

# adding relevant columns
negative_df_results["mirna_sequence"] = negative_df["mirna_sequence"]
negative_df_results["mre_region"] = negative_df["mrna_sequence"]
negative_df_results["mirna_accession"] = negative_df["mirna_accession"]
negative_df_results["enst"] = clash_df["enst"]



# export to csv
negative_df_results.to_csv("results/3_negative_data.csv", index=False)
