==> notebooks/4_merge_data.ipynb <==
%cd ..

import pandas as pd

def generate_alignment_strings(df):

    full_strings = []

    for _, row in df.iterrows():

        start_string = (row.mirna_start - 1) * "0"
        mid_string = row["mirna_dot_bracket_5to3"].replace(
            ".", "0").replace(")", "1")
        end_string = (len(row.mirna_sequence) - row.mirna_end) * "0"

        full_string = start_string + mid_string + end_string

        full_strings.append(full_string)

    df["alignment_string"] = full_strings

    return df

def count_basepairs(df):

    def count_ones(str, seed=False):
        return str[1:7].count("1") if seed else str.count("1")

    df["pred_num_basepairs"] = df["alignment_string"].apply(count_ones)

    df["pred_seed_basepairs"] = df["alignment_string"].apply(
        count_ones, seed=True)

    return df

positive_df = pd.read_csv("results/2_positive_data.csv")
positive_df["label"] = 1


negative_df = pd.read_csv("results/3_negative_data.csv")
negative_df["label"] = 0

df = pd.concat([positive_df, negative_df], ignore_index=True)
df = generate_alignment_strings(df)
df = count_basepairs(df)

df.to_csv("results/4_merged_data.csv", index=False)
