==> notebooks/7_process_vcfs.ipynb <==
cd ..

import pandas as pd
from Bio import SeqIO

df = pd.read_csv("data/sample_vcfs/PD3851a.vcf", sep="\t", header=None)
df.columns = ["chr", "pos", "unknown", "ref", "alt"]
df.head()

len(df)

from Bio import SeqIO

def read_fasta(file_path):
    """
    Reads a FASTA file using SeqIO from Biopython.

    Args:
        file_path (str): Path to the FASTA file.

    Returns:
        list: List of SeqRecord objects representing the sequences in the FASTA file.
    """
    sequences = list(SeqIO.parse(file_path, "fasta"))
    return sequences

asd = read_fasta("data/hg38/Homo_sapiens.GRCh38.dna.chromosome.1.fa")



for record in SeqIO.parse("data/hg38/Homo_sapiens.GRCh38.dna.chromosome.1.fa", "fasta"):
    print(record.id)

record = next(SeqIO.parse("data/hg38/Homo_sapiens.GRCh38.dna.chromosome.1.fa", 'fasta'))

record.seq[1]

asd[0].seq

from Bio import SeqIO
def get_nucleotide_at_position_seqio(chrom, position):

    sequences = SeqIO.parse(f"data/hg38/Homo_sapiens.GRCh38.dna.chromosome.{chrom}.fa", "fasta")
    sequence = next(sequences, None)
    if not sequence or position < 0 or position >= len(sequence):
        return None  # Invalid position or empty sequence

    return sequence.seq[position]


get_nucleotide_at_position_seqio(1, 1911010)

def get_nucleotide_at_position(chrom, position):
    file_path = f"data/hg38/Homo_sapiens.GRCh38.dna.chromosome.{chrom}.fa"
    with open(file_path, 'r') as file:
        file.readline()
        byte_position = file.tell()
        line_length = len(file.readline().strip())
        offset = position - 1
        num_new_lines = offset // line_length
        byte_position = byte_position + offset + num_new_lines
        file.seek(byte_position)

        # Read the nucleotide at the position
        nucleotide = file.read(1)
    return nucleotide
get_nucleotide_at_position(1, 100002)



chrom = 1
file_path = f"data/hg38/Homo_sapiens.GRCh38.dna.chromosome.{chrom}.fa"
with open(file_path, 'r') as file:
    file.seek(2)
    
    asd =file.read(1)
    print(asd)
